Upgrade zetasql to 2022.02.1
This upgrades zetasql to the latest version to pull in some updates to
the build tooling. While the previous version compiled, it had an older
version of the build toolchain which caused the build to not be
optimized. This resulted in a working binary but one that was ~30%
slower than running the emlated amd64 docker container. Upgrading to the
latest zetasql improved performance of the arm64 binary by ~155% based
on local testing.